added Dashboard and some fixes

+++ backend/src/utils/cloudinary.js
@@ -15,7 +15,9 @@ const uploadOnCloudinary=async (localFilePath)=>{
         const response = await cloudinary.uploader.upload(localFilePath,{
             resource_type:"auto"
         })
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath)
+        }
 
         return response;
     }catch(error){
